// Capstone Project/src/app/pages/admin/admin-account-management-page/screen/types.ts
export enum adminPageAccountEnum {
  GET_DATA_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/GET_DATA_ACCOUNT_REQUEST',
  GET_DATA_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/GET_DATA_ACCOUNT_SUCCESS',
  GET_DATA_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/GET_DATA_ACCOUNT_FAILURE',

  UPDATE_DATA_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/UPDATE_DATA_ACCOUNT_REQUEST',
  UPDATE_DATA_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_DATA_ACCOUNT_SUCCESS',
  UPDATE_DATA_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/UPDATE_DATA_ACCOUNT_FAILURE',

  CREATE_DATA_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_ACCOUNT_REQUEST',
  CREATE_DATA_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/CREATE_DATA_ACCOUNT_SUCCESS',
  CREATE_DATA_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/CREATE_DATA_ACCOUNT_FAILURE',

  DELETE_DATA_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/DELETE_DATA_ACCOUNT_REQUEST',
  DELETE_DATA_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/DELETE_DATA_ACCOUNT_SUCCESS',
  DELETE_DATA_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/DELETE_DATA_ACCOUNT_FAILURE',

  LOCK_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/LOCK_ACCOUNT_REQUEST',
  LOCK_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/LOCK_ACCOUNT_SUCCESS',
  LOCK_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/LOCK_ACCOUNT_FAILURE',

  UNLOCK_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/UNLOCK_ACCOUNT_REQUEST',
  UNLOCK_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/UNLOCK_ACCOUNT_SUCCESS',
  UNLOCK_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/UNLOCK_ACCOUNT_FAILURE',

  CREATE_NEW_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/CREATE_NEW_ACCOUNT_REQUEST',
  CREATE_NEW_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/CREATE_NEW_ACCOUNT_SUCCESS',
  CREATE_NEW_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/CREATE_NEW_ACCOUNT_FAILURE',


  VERIFY_ACCOUNT_REQUEST = 'APP/ADMIN_PAGE/VERIFY_ACCOUNT_REQUEST',
  VERIFY_ACCOUNT_SUCCESS = 'APP/ADMIN_PAGE/VERIFY_ACCOUNT_SUCCESS',
  VERIFY_ACCOUNT_FAILURE = 'APP/ADMIN_PAGE/VERIFY_ACCOUNT_FAILURE',

  CLEAR_DATA_ACCOUNT = 'APP/ADMIN_PAGE/CLEAR_DATA_ACCOUNT',
}